document_safe/document_normalizer/requirements.txt
 aspose-pdf